Moved drill grid explorer into the ratings section
#2896
