+++ waltz-ng/client/measurable-rating/components/browser-section/measurable-ratings-browser-section.js
@@ -19,9 +19,6 @@
 
 import _ from 'lodash';
 import {initialiseData} from '../../../common';
-import {mkLinkGridCell} from '../../../common/link-utils';
-import {mkSelectionOptions} from '../../../common/selector-utils';
-import {CORE_API} from '../../../common/services/core-api-utils';
 import template from './measurable-ratings-browser-section.html';
 
 /**
@@ -38,174 +35,31 @@ const bindings = {
 
 
 const initialState = {
-    applications: [],
-    measurables: [],
-    measurableCategories: [],
-    measurableRatingsDetail: null,
-    ratingTallies: [],
-    detail: null,
     visibility: {
-        loading: false
-    },
-    selectedMeasurable: null,
-    onLoadDetail: () => log('onLoadDetail')
+        treeView: true,
+        gridView: false
+    }
 };
 
 
-function prepareColumnDefs(measurableCategory) {
-    return [
-        mkLinkGridCell('Name', 'application.name', 'application.id', 'main.app.view'),
-        {
-            field: 'application.assetCode',
-            name: 'Asset Code'
-        },
-        {
-            field: 'ratingName',
-            name: 'Rating',
-            cellTemplate: '<div class="ui-grid-cell-contents"><waltz-rating-indicator-cell rating="row.entity.rating.rating" label="COL_FIELD.name"></waltz-rating-indicator-cell></div>'
-        },
-        {
-            field: 'measurable.name',
-            name: measurableCategory.name
-        },
-        {
-            field: 'rating.description',
-            name: 'Comment'
-        }
-    ];
-}
 
-
-function findChildIds(measurable) {
-    const recurse = (acc, n) => {
-        acc.push(n.id);
-        _.each(n.children, c => recurse(acc, c));
-        return acc;
-    };
-
-    return recurse([], measurable);
-}
-
-
-function prepareTableData(measurable,
-                          ragNames = {},
-                          applications = [],
-                          ratings = [],
-                          measurablesById = {}) {
-
-    const appsById = _.keyBy(applications, 'id');
-    const relevantMeasurableIds = findChildIds(measurable);
-
-    const data = _
-        .chain(ratings)
-        .filter(r => _.includes(relevantMeasurableIds, r.measurableId))
-        .map(r => {
-            return {
-                application: appsById[r.entityReference.id],
-                rating: r,
-                measurable: measurablesById[r.measurableId],
-                ratingName: ragNames[r.rating]
-            };
-        })
-        .value();
-
-    return _.sortBy(data, 'application.name');
-}
-
-
-function log() {
-    console.log("wmrbs::", arguments);
-}
-
-
-function loadMeasurableRatings(serviceBroker, selector, holder) {
-    return serviceBroker
-        .loadViewData(CORE_API.MeasurableRatingStore.statsByAppSelector, [selector])
-        .then(r => holder.ratingTallies = r.data);
-}
-
-
-function loadMeasurables(serviceBroker, selector, holder) {
-    return serviceBroker
-        .loadViewData(CORE_API.MeasurableStore.findMeasurablesBySelector, [selector])
-        .then(r => holder.measurables = r.data);
-}
-
-
-function loadMeasurableCategories(serviceBroker, holder) {
-    return serviceBroker
-        .loadAppData(CORE_API.MeasurableCategoryStore.findAll, [])
-        .then(r => holder.measurableCategories = r.data);
-}
-
-
-function loadApps(serviceBroker, selector, holder) {
-    return serviceBroker
-        .loadViewData(CORE_API.ApplicationStore.findBySelector, [selector])
-        .then(r => holder.applications = r.data);
-}
-
-
-function controller($q, serviceBroker) {
+function controller() {
     const vm = initialiseData(this, initialState);
 
-    const loadRatingsDetail = () => {
-        return vm.measurableRatingsDetail
-            ? $q.resolve(vm.measurableRatingsDetail)
-            : serviceBroker
-                .execute(CORE_API.MeasurableRatingStore.findByAppSelector, [vm.selector])
-                .then(r => vm.measurableRatingsDetail = r.data);
+
+    vm.showGridView = () => {
+        vm.visibility.gridView = true;
+        vm.visibility.treeView = false;
     };
 
-
-    vm.$onInit = () => {
-        vm.selector = mkSelectionOptions(vm.parentEntityRef);
-
-        return $q.all([
-            loadMeasurableRatings(serviceBroker, vm.selector, vm),
-            loadMeasurables(serviceBroker, vm.selector, vm),
-            loadMeasurableCategories(serviceBroker, vm),
-            loadApps(serviceBroker, vm.selector, vm),
-        ]);
+    vm.showTreeView = () => {
+        vm.visibility.gridView = false;
+        vm.visibility.treeView = true;
     };
-
-
-    vm.$onChanges = (c) => {
-        if (c.measurables) {
-            vm.measurablesById = _.keyBy(vm.measurables, 'id');
-        }
-    };
-
-
-    vm.onSelect = (measurable) => {
-        vm.visibility.loading = true;
-        vm.tableData = null;
-        vm.selectedMeasurable = measurable;
-        const promise = loadRatingsDetail();
-        const category = _.find(vm.measurableCategories, ({ id: measurable.categoryId }));
-
-        vm.columnDefs = prepareColumnDefs(category);
-        if (_.isFunction(_.get(promise, 'then'))) {
-            promise
-                .then(ratings => vm.tableData = prepareTableData(
-                    measurable,
-                    category.ragNames,
-                    vm.applications,
-                    ratings,
-                    vm.measurablesById))
-                .then(() => vm.visibility.loading = false);
-        } else {
-            log('was expecting promise, got: ', promise);
-            vm.visibility.loading = false;
-        }
-    };
-
 }
 
 
 controller.$inject = [
-    '$q',
-    'ServiceBroker'
 ];
